test/requirements.txt
opencensus-ext-azure==1.1.0
PyYAML==6.0
